Init Commit: organismLife and exec for macos
organismLife.py is used to illustrate that organisms in
principle can be understood as purely mechanistic systems.
The organisms simulated do not have a central homunculus
as it were, and their visual experience is simple.

My point here is to question what it means to have qualia.
Where in the brain does the buck stop when it comes to experience?
I have qualia and experience, but who is it that has it? For
whom is the experience?

Dennett talks about the cartesian theater, and I think he is
right to do so. But I think he is wrong to say that there is
no homunculus. I think there is a homunculus, but it is not
a central homunculus. It is a distributed homunculus maybe.

I do not know. But I am thinking here about some kind of unity
or experience, Kant calls it the transcendental unity of apperception.
It is the unity of experience, the I that has experience, that
accompanies all experience.